Create separate navigation for measurement and workout
