+++ composeApp/src/commonMain/kotlin/io/github/jakubherr/gitfit/domain/Util.kt
@@ -11,7 +11,4 @@ fun Double?.isNonNegative() = this != null && this >= 0.0
 fun String.isPositiveInt() = toIntOrNull().let { it != null && it > 0 }
 fun String.isPositiveDouble() = toDoubleOrNull().let { it != null && it > 0 }
 
-// note: this function is for UI purposes only and is double-checked by backend
-fun String.isStrongPassword() = isNotBlank() && length >= 12
-
 fun String.validDecimals(maxDecimals: Int) = substringAfter(".", "").length <= maxDecimals
