+++ composeApp/src/commonMain/kotlin/io/github/jakubherr/gitfit/presentation/workout/LoggedInGraph.kt
@@ -1,25 +1,11 @@
 package io.github.jakubherr.gitfit.presentation.workout
 
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.composable
 import androidx.navigation.navigation
-import gitfit.composeapp.generated.resources.Res
-import gitfit.composeapp.generated.resources.empty_workout_history
-import gitfit.composeapp.generated.resources.error_block_with_progression
-import gitfit.composeapp.generated.resources.error_workout_not_found
-import io.github.jakubherr.gitfit.presentation.AddExerciseToWorkoutRoute
 import io.github.jakubherr.gitfit.presentation.DashboardRoute
 import io.github.jakubherr.gitfit.presentation.ExerciseListRoute
 import io.github.jakubherr.gitfit.presentation.HistoryRoute
@@ -27,7 +13,6 @@ import io.github.jakubherr.gitfit.presentation.LoggedInRoute
 import io.github.jakubherr.gitfit.presentation.MeasurementHistoryRoute
 import io.github.jakubherr.gitfit.presentation.PlanDetailRoute
 import io.github.jakubherr.gitfit.presentation.SettingsRoute
-import io.github.jakubherr.gitfit.presentation.WorkoutDetailRoute
 import io.github.jakubherr.gitfit.presentation.WorkoutHistoryRoute
 import io.github.jakubherr.gitfit.presentation.WorkoutInProgressRoute
 import io.github.jakubherr.gitfit.presentation.auth.AuthViewModel
@@ -35,16 +20,14 @@ import io.github.jakubherr.gitfit.presentation.dashboard.DashboardAction
 import io.github.jakubherr.gitfit.presentation.dashboard.DashboardScreenRoot
 import io.github.jakubherr.gitfit.presentation.exercise.exerciseNavigation
 import io.github.jakubherr.gitfit.presentation.graph.HistoryScreenRoot
-import io.github.jakubherr.gitfit.presentation.measurement.measurementGraph
+import io.github.jakubherr.gitfit.presentation.measurement.measurementNavigation
 import io.github.jakubherr.gitfit.presentation.planning.PlanningViewModel
-import io.github.jakubherr.gitfit.presentation.planning.planningGraph
+import io.github.jakubherr.gitfit.presentation.planning.planningNavigation
 import io.github.jakubherr.gitfit.presentation.settings.SettingsScreenRoot
-import kotlinx.coroutines.launch
-import org.jetbrains.compose.resources.getString
-import org.jetbrains.compose.resources.stringResource
 import org.koin.compose.viewmodel.koinNavViewModel
 import org.koin.core.annotation.KoinExperimentalAPI
 
+// This nested navigation graph contains all destinations that require a logged in user
 fun NavGraphBuilder.loggedInGraph(
     navController: NavHostController,
     showSnackbar: (String) -> Unit,
@@ -76,50 +59,13 @@ fun NavGraphBuilder.loggedInGraph(
             )
         }
 
-        composable<WorkoutInProgressRoute> {
-            val vm = navController.sharedViewModel<WorkoutViewModel>()
-            val scope = rememberCoroutineScope()
+        workoutNavigation(navController, showSnackbar = showSnackbar)
 
-            WorkoutInProgressScreenRoot(
-                vm = vm,
-                onAction = { action ->
-                    if (action !is WorkoutAction.CompleteCurrentWorkout && action !is WorkoutAction.RemoveBlock) vm.onAction(
-                        action
-                    )
+        planningNavigation(navController, showSnackbar = showSnackbar)
 
-                    when (action) {
-                        is WorkoutAction.AskForExercise -> navController.navigate(AddExerciseToWorkoutRoute(action.workoutId))
-                        is WorkoutAction.DeleteWorkout -> navController.popBackStack()
-                        is WorkoutAction.CompleteCurrentWorkout -> {
-                            val error = vm.currentWorkout.value?.error
-                            if (error == null) {
-                                vm.onAction(action)
-                            } else {
-                                showSnackbar(error.message)
-                            }
-                        }
-                        is WorkoutAction.RemoveBlock -> {
-                            if (action.block.progressionSettings != null) {
-                                scope.launch { showSnackbar(getString(Res.string.error_block_with_progression)) }
-                            } else {
-                                vm.onAction(action)
-                            }
-                        }
-                        else -> {}
-                    }
-                },
-                onSaveComplete = {
-                    vm.onAction(WorkoutAction.NotifyWorkoutSaved)
-                    navController.popBackStack()
-                },
-            )
-        }
+        measurementNavigation(navController, showSnackbar = showSnackbar)
 
         exerciseNavigation(navController)
-
-        measurementGraph(navController, showSnackbar = showSnackbar)
-
-        planningGraph(navController, showSnackbar = showSnackbar)
 
         composable<HistoryRoute> {
             HistoryScreenRoot(
@@ -129,46 +75,6 @@ fun NavGraphBuilder.loggedInGraph(
                 onBrowseExerciseData = { navController.navigate(ExerciseListRoute) },
                 onBrowseMeasurementData = { navController.navigate(MeasurementHistoryRoute) },
             )
-        }
-
-        composable<WorkoutHistoryRoute> {
-            val vm = navController.sharedViewModel<WorkoutViewModel>()
-            val completedWorkouts by vm.completedWorkouts.collectAsStateWithLifecycle()
-
-            if (completedWorkouts.isEmpty()) {
-                Column(
-                    Modifier.fillMaxSize(),
-                    verticalArrangement = Arrangement.Center,
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                ) {
-                    Text(stringResource(Res.string.empty_workout_history))
-                }
-            } else {
-                WorkoutListScreen(
-                    completedWorkouts,
-                    onWorkoutSelected = {
-                        vm.onAction(WorkoutAction.SelectWorkout(it))
-                        navController.navigate(WorkoutDetailRoute)
-                    },
-                )
-            }
-        }
-
-        composable<WorkoutDetailRoute> {
-            val vm = navController.sharedViewModel<WorkoutViewModel>()
-            val workout = vm.selectedWorkout
-            workout?.let {
-                WorkoutDetailScreen(
-                    it,
-                    onDelete = {
-                        vm.onAction(WorkoutAction.DeleteWorkout(it.id))
-                        navController.popBackStack()
-                    },
-                )
-            }
-            if (workout == null) {
-                Text(stringResource(Res.string.error_workout_not_found))
-            }
         }
 
         composable<SettingsRoute> {
